refactor: 데이터베이스 객체로 상품 정보 추출
